feat(Recorder): Add create test menu to make exporting browser tests easier
